Cypress: Separate dashboard spec
